+++ environment/providers/secretdir/secretdir.go
@@ -11,14 +11,18 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/nutanix-cloud-native/prism-go-client/environment/credentials"
 	"github.com/nutanix-cloud-native/prism-go-client/environment/types"
 )
 
 const (
-	secretKeyName         = "key"
-	secretKeyCertName     = "cert"
-	secretKeyCertEndpoint = "endpoint"
-	envSecretDir          = "CSI_SECRET_DIR"
+	secretKeyName        = "key"
+	secretKeyCertName    = "cert"
+	secretKeyEndpoint    = "endpoint"
+	secretKeyInsecure    = "insecure"
+	secretKeyCredentials = credentials.KeyName
+	envCSISecretDir      = "CSI_SECRET_DIR"
+	envSecretDir         = "SECRET_DIR"
 )
 
 var ErrMissingValue = fmt.Errorf("missing information in secret value " +
@@ -32,7 +36,12 @@ func (prov *provider) GetManagementEndpoint(
 ) (*types.ManagementEndpoint, error) {
 	path := os.Getenv(envSecretDir)
 	if path == "" {
-		return nil, fmt.Errorf("environment variable %s not set", envSecretDir)
+		// fall back to CSI env var
+		path = os.Getenv(envCSISecretDir)
+		if path == "" {
+			return nil, fmt.Errorf("neither %s nor %s environment variable set",
+				envSecretDir, envCSISecretDir)
+		}
 	}
 	return getMgmtEndpointFromSecretDir(path)
 }
@@ -59,25 +68,71 @@ func readParam(path, key string) (string, error) {
 	}
 }
 
+func getApiEndpoint(path string) (*url.URL, error) {
+	endpoint, err := readParam(path, secretKeyEndpoint)
+	if err != nil {
+		return nil, err
+	}
+	if endpoint == "" {
+		return nil, fmt.Errorf("no endpoint is provided in secret value")
+	}
+	return url.Parse(fmt.Sprintf("https://%s", endpoint))
+}
+
+func getInsecure(path string) (bool, error) {
+	val, err := readParam(path, secretKeyInsecure)
+	if err != nil {
+		return false, err
+	}
+	if val == "" {
+		return false, nil
+	}
+	switch val {
+	case "True", "true", "1", "on", "yes":
+		return true, nil
+	default:
+		return false, nil
+	}
+}
+
 // mangementEndpointFromSecretParams parses management endpoint and credentials from
 // parameters embedded as data in a k8s secret.
 func getMgmtEndpointFromSecretDir(path string) (
 	*types.ManagementEndpoint, error,
 ) {
+	// Auto-detect new converged secret format
+	credsData, err := readParam(path, secretKeyCredentials)
+	if err != nil {
+		return nil, err
+	}
+	if credsData != "" {
+		apiCreds, err := credentials.ParseCredentials([]byte(credsData))
+		if err != nil {
+			return nil, err
+		}
+		addr, err := getApiEndpoint(path)
+		if err != nil {
+			return nil, err
+		}
+		insecure, err := getInsecure(path)
+		if err != nil {
+			return nil, err
+		}
+		return &types.ManagementEndpoint{
+			Address:        addr,
+			Insecure:       insecure,
+			ApiCredentials: *apiCreds,
+		}, nil
+	}
+
+	// Handle CSI format
 	certString, err := readParam(path, secretKeyCertName)
 	if err != nil {
 		return nil, err
 	}
 	if certString != "" {
 		// TLS key pair
-		endpoint, err := readParam(path, secretKeyCertEndpoint)
-		if err != nil {
-			return nil, err
-		}
-		if endpoint == "" {
-			return nil, fmt.Errorf("no endpoint is provided in secret value")
-		}
-		addr, err := url.Parse(fmt.Sprintf("https://%s", endpoint))
+		addr, err := getApiEndpoint(path)
 		if err != nil {
 			return nil, err
 		}